VJ-26: Реализовать получение списка услуг по фильтрам
Фикс 1 - Не видело конструктор

+++ src/main/java/com/vpr42/marketplacefeedapi/mappers/JobsMapper.java
@@ -16,7 +16,7 @@ import java.util.stream.Collectors;
  */
 @Slf4j
 public class JobsMapper {
-    public static Job fromEntity(JobEntity entity, int orderCount) {
+    public static Job fromEntity(JobEntity entity, long orderCount) {
         log.info("Converting ServiceEntity to dto: {}", entity);
 
         return new Job(
