+++ src/main/java/com/vpr42/marketplacefeedapi/model/dto/JobEntityWithCount.java
@@ -4,5 +4,7 @@ import com.vpr42.marketplacefeedapi.model.entity.JobEntity;
 
 public record JobEntityWithCount(
     JobEntity job,
-    Integer count
-) { }
+    Long count
+) {
+
+}
